requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
psycopg[binary]
whitenoise==6.7.0
Pillow==10.4.0
gunicorn==22.0.0
python-dotenv==1.0.1
